requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
python-socketio==5.4.0

transformers==4.11.3
tokenizers==0.10.3

torch==1.12.1
numpy==1.26.4
scikit-learn==1.4.1.post1

pytesseract==0.3.8
Pillow==8.3.2
PyMuPDF==1.18.17

python-engineio==4.2.1

gevent==23.9.1
gevent-websocket==0.10.1
greenlet==2.0.2
